feat: add exception handling to all services

+++ src/main/java/com/wise/dental_clinic/services/PatientService.java
@@ -3,10 +3,15 @@ package com.wise.dental_clinic.services;
 import com.wise.dental_clinic.dto.PatientDTO;
 import com.wise.dental_clinic.entities.Patient;
 import com.wise.dental_clinic.repositories.PatientRepository;
+import com.wise.dental_clinic.services.exceptions.DatabaseException;
+import com.wise.dental_clinic.services.exceptions.ResourceNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @Service
@@ -27,7 +32,7 @@ public class PatientService {
     @Transactional(readOnly = true)
     public PatientDTO findById(Long id) {
         Optional<Patient> result = repository.findById(id);
-        Patient patient = result.orElseThrow();
+        Patient patient = result.orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
         return new PatientDTO(patient);
     }
 
@@ -40,15 +45,26 @@ public class PatientService {
 
     @Transactional
     public PatientDTO update(PatientDTO dto, Long id) {
-        Optional<Patient> result = repository.findById(id);
-        Patient entity = result.orElseThrow();
-        entityToDto(entity, dto);
-        return new PatientDTO(entity);
+        try {
+            Optional<Patient> result = repository.findById(id);
+            Patient entity = result.orElseThrow();
+            entityToDto(entity, dto);
+            return new PatientDTO(entity);
+        } catch (NoSuchElementException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void entityToDto(Patient entity, PatientDTO dto) {
